--- components/panel/examples/_utils.js
@@ -1,13 +1,0 @@
-/** @jsx jsx */
-
-import { jsx } from '@westpac/core';
-
-export const Box = props => (
-	<div
-		css={{
-			alignItems: 'center',
-			display: 'flex',
-		}}
-		{...props}
-	/>
-);
