+++ components/panel/src/index.js
@@ -1,1 +1,4 @@
-export { Panel, PanelHeader, PanelBody, PanelFooter } from './Panel';
+export { Panel } from './Panel';
+export { PanelHeader } from './PanelHeader';
+export { PanelBody } from './PanelBody';
+export { PanelFooter } from './PanelFooter';
